Updated mypy configuration for versions >=0.600
With version 0.600 the way mypy uses per-module configuration has changed (see https://github.com/python/mypy/pull/4894) and the glob pattern 'test_*' could not be used to exclude tests. The solution was to add a '__init__.py' file to identify 'tests' as a proper package and exclude it completely from mypy checks.